fix: move `HuffmanDecoder` to `compressors.support`
